kvm: qemu: migration: save/load PCI-Bus irq related state.
This irq state is being used to decide whether or not an interrupt
should be passed to the guest or not.

This state can be calculated from all the pci devices on that bus.
This commit may be reverted if we decide to calculate the irq related state
instead of save/load it.

Signed-off-by: Uri Lublin <uril@qumranet.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>