kvm: qemu: virtio block device
This patch implements the backend support for the virtio block device.  It's
designed to support in-order queueing of a virtually unlimited size so it will
be able to perform better than SCSI.  Besides performance, the virtio block
interface passes through guest SCSI commands so it can be used to expose any
type of block device (although only normal disks are supported in this patch).

[dor: use Qumranet donated pci ids instead of unregistered pci ids]

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Dor Laor <dor.laor@qumranet.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>