kvm: qemu: save 64-bit of the IA-32e capable sysenter MSRs
When transitioning from KVM to the qemu userspace, we try to get and
push a whole bunch of MSR values, including the SYSENTER ones. While
this is basically a good idea, qemu doesn't know anything about SYSENTER
on x86_64, which is an additional feature only available on Intel CPUs
and thus defines SYSENTER_EIP and SYSENTER_ESP as 32-bit values.

Because we're saving/restoring the SYSENTER values on transitions to
userspace and the values in env->cpu only hold 32 bits, they get
truncated to 32-bit. The easiest way around this is to widen the fields
to 64 bits. This should not disturb the current functionality, but
allows us to run 32-bit code on x86_64 using VMX.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Avi Kivity <avi@qumranet.com>