kvm: qemu: Optionally bounce the virtio buffers
In the absence of a zero-copy API, we cannot touch physical memory directly.
This patch introduces a #define to control whether we touch buffers directly,
or bounce the buffers through cpu_physical_memory_rw.

The #define is needed for merging virtio into upstream QEMU since zero-copy will
not be accepted as-is.  As a bonus, I also added physical memory dirtying
support to the zero-copy case so now live migration should work when using
virtio.

With this patch, virtio is mergable into upstream QEMU (except for virtio-net,
that's a different series though).

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>